Add CPC seasonal outlooks, CDEC snow stations, and skip_existing option

+++ data_download/wsfr_download/mjo.py
@@ -13,8 +13,11 @@ See the challenge website for more about this approved data source:
 https://www.drivendata.org/competitions/254/reclamation-water-supply-forecast-dev/page/801/#madden-julian-oscillation-mjo-pentad-indices
 """
 
+from typing import Annotated
+
 from loguru import logger
 import requests
+import typer
 
 from wsfr_download.config import DATA_ROOT
 
@@ -24,12 +27,19 @@ SOURCE_URL = (
 FILE_PATH_PARTS = ("teleconnections", "mjo.txt")
 
 
-def download_mjo():
+def download_mjo(
+    skip_existing: Annotated[bool, typer.Option(help="Whether to skip an existing file.")] = True,
+):
     """Download Madden-Julian Oscillation indices."""
     logger.info("Downloading MJO data...")
     response = requests.get(SOURCE_URL)
     out_file = DATA_ROOT.joinpath(*FILE_PATH_PARTS)
-    out_file.parent.mkdir(exist_ok=True, parents=True)
-    with out_file.open("w") as fp:
-        fp.write(response.text)
-    logger.success(f"MJO data written to {out_file}")
+    logger.info(f"Output file path is {out_file}")
+    if skip_existing and out_file.exists():
+        logger.info("File exists. Skipping.")
+    else:
+        out_file.parent.mkdir(exist_ok=True, parents=True)
+        with out_file.open("w") as fp:
+            fp.write(response.text)
+        logger.info("Data downloaded to file.")
+    logger.success("MJO download complete.")
