+++ data_download/wsfr_download/soi.py
@@ -12,8 +12,11 @@ See the challenge website for more about this approved data source:
 https://www.drivendata.org/competitions/254/reclamation-water-supply-forecast-dev/page/801/#southern-oscillation-index-soi
 """
 
+from typing import Annotated
+
 from loguru import logger
 import requests
+import typer
 
 from wsfr_download.config import DATA_ROOT
 
@@ -21,12 +24,19 @@ SOURCE_URL = "https://www.cpc.ncep.noaa.gov/data/indices/soi"
 FILE_PATH_PARTS = ("teleconnections", "soi.txt")
 
 
-def download_soi():
+def download_soi(
+    skip_existing: Annotated[bool, typer.Option(help="Whether to skip an existing file.")] = True,
+):
     """Download Southern Oscillation Index data."""
     logger.info("Downloading SOI data...")
     response = requests.get(SOURCE_URL)
     out_file = DATA_ROOT.joinpath(*FILE_PATH_PARTS)
-    out_file.parent.mkdir(exist_ok=True, parents=True)
-    with out_file.open("w") as fp:
-        fp.write(response.text)
-    logger.success(f"SOI data written to {out_file}")
+    logger.info(f"Output file path is {out_file}")
+    if skip_existing and out_file.exists():
+        logger.info("File exists. Skipping.")
+    else:
+        out_file.parent.mkdir(exist_ok=True, parents=True)
+        with out_file.open("w") as fp:
+            fp.write(response.text)
+        logger.info("Data downloaded to file.")
+    logger.success("SOI download complete.")
